Fixed return type

+++ include/battle.h
@@ -6,7 +6,7 @@ double getTypeEffect(Move move, Type type);
 
 int getDamage(Move move, Pokemon poke1, Pokemon poke2, bool crit, double roll);
 
-int getDamage(Move move, Pokemon poke1, Pokemon poke2, bool STAB);
+double getDamage(Move move, Pokemon poke1, Pokemon poke2, bool STAB);
 
 bool isPhysical(Move move);
